# docs/openapi.yml
...
openapi: 3.1.0
paths:
...
  /artifacts/api/gc/plan:
    post:
      operationId: planGcWithSuppliedPins
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GcPlanReport"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Plan With Supplied Pins
      tags:
      - Gc Plan Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:system
    get:
      operationId: getGcPlan
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GcPlanReport"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Plan
      tags:
      - Gc Plan Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
        - qits:system
        - qits:ci-run
  /artifacts/api/gc/sweep:
    post:
      operationId: sweepGc
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GcSweepReport"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Sweep
      tags:
      - Gc Plan Controller
      security:
      - SecurityScheme:
        - qits:system
  /artifacts/api/repositories/{repo}/daemons/{daemon}/versions:
    get:
      operationId: listDaemonVersions
      parameters:
      - name: daemon
        in: path
        required: true
        schema:
          type: string
      - name: repo
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDaemonVersionsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Versions
      tags:
      - Daemon Browse Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
        - qits:system
        - qits:ci-run
...
components:
  schemas:
    DaemonVersionSummary:
      type: object
      properties:
        version:
          type: string
        digest:
          type: string
        sizeBytes:
          type: integer
          format: int64
        publishedAt:
          $ref: "#/components/schemas/Instant"
        accessedAt:
          $ref: "#/components/schemas/Instant"
    GcIdentity:
      type: object
      properties:
        repository:
          type: string
        identity:
          type: string
        rule:
          type: string
    GcPinSource:
      type: object
      properties:
        source:
          type: string
        url:
          type: string
        answered:
          type: boolean
        outcome:
          type: string
        readAt:
          $ref: "#/components/schemas/Instant"
        tookMillis:
          type: integer
          format: int64
        pinCount:
          type: integer
          format: int32
        keeps:
          type: array
          items:
            type: string
    GcPlanReport:
      type: object
      properties:
        summary:
          $ref: "#/components/schemas/GcPlanSummary"
        generatedAt:
          $ref: "#/components/schemas/Instant"
        dryRun:
          type: boolean
        graceWindow:
          type: string
        executable:
          type: boolean
        pinFailures:
          type: array
          items:
            type: string
        pins:
          type: array
          items:
            $ref: "#/components/schemas/GcPinSource"
        configuration:
          type: array
          items:
            $ref: "#/components/schemas/GcTypeConfiguration"
        types:
          type: array
          items:
            $ref: "#/components/schemas/GcTypePlan"
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/GcRepositoryPlanSummary"
        sweep:
          $ref: "#/components/schemas/GcSweepPlan"
        untouchable:
          $ref: "#/components/schemas/GcUntouchablePool"
    GcPlanSummary:
      type: object
      properties:
        executable:
          type: boolean
        headline:
          type: string
        identitiesCondemned:
          type: integer
          format: int32
        blobsSweepable:
          type: integer
          format: int32
        reclaimableBytes:
          type: integer
          format: int64
        reclaimable:
          type: string
        withheldByGraceWindow:
          type: integer
          format: int32
        types:
          type: array
          items:
            type: string
    GcRepositoryPlanSummary:
      type: object
      properties:
        repository:
          type: string
        type:
          type: string
        strategy:
          type: string
        note:
          type: string
        error:
          type: string
        identitiesCondemned:
          type: integer
          format: int32
        identitiesKept:
          type: integer
          format: int32
        blobsSweepable:
          type: integer
          format: int32
        reclaimableBytes:
          type: integer
          format: int64
        withheldByGraceWindow:
          type: integer
          format: int32
        withheldBytes:
          type: integer
          format: int64
    GcSweepOutcome:
      type: object
      properties:
        blobsUnlinked:
          type: integer
          format: int32
        bytesReclaimed:
          type: integer
          format: int64
        withheldByGraceWindow:
          type: integer
          format: int32
        withheldBytes:
          type: integer
          format: int64
        stillReferenced:
          type: integer
          format: int32
        alreadyGone:
          type: integer
          format: int32
        unlinkedBlobIds:
          type: array
          items:
            type: string
    GcSweepPlan:
      type: object
      properties:
        blobCount:
          type: integer
          format: int32
        reclaimableBytes:
          type: integer
          format: int64
        withheldByGraceWindow:
          type: integer
          format: int32
        withheldBytes:
          type: integer
          format: int64
        blobIds:
          type: array
          items:
            type: string
    GcSweepReport:
      type: object
      properties:
        executedAt:
          $ref: "#/components/schemas/Instant"
        dryRun:
          type: boolean
        graceWindow:
          type: string
        aborted:
          type: string
        pins:
          type: array
          items:
            $ref: "#/components/schemas/GcPinSource"
        types:
          type: array
          items:
            $ref: "#/components/schemas/GcTypeSweepResult"
        sweep:
          $ref: "#/components/schemas/GcSweepOutcome"
        untouchable:
          $ref: "#/components/schemas/GcUntouchablePool"
    GcTypeConfiguration:
      type: object
      properties:
        type:
          type: string
        strategy:
          type: string
        window:
          type: string
        rule:
          type: string
    GcTypePlan:
      type: object
      properties:
        type:
          type: string
        strategy:
          type: string
        note:
          type: string
        error:
          type: string
        dead:
          type: array
          items:
            $ref: "#/components/schemas/GcIdentity"
        kept:
          type: array
          items:
            $ref: "#/components/schemas/GcIdentity"
        blobsReleased:
          type: integer
          format: int32
        blobsSweepable:
          type: integer
          format: int32
        reclaimableBytes:
          type: integer
          format: int64
    GcTypeSweepResult:
      type: object
      properties:
        type:
          type: string
        strategy:
          type: string
        note:
          type: string
        error:
          type: string
        deleted:
          type: array
          items:
            $ref: "#/components/schemas/GcIdentity"
        withheldByGraceWindow:
          type: array
          items:
            $ref: "#/components/schemas/GcIdentity"
    GcUntouchablePool:
      type: object
      properties:
        reason:
          type: string
        blobCount:
          type: integer
          format: int32
        bytes:
          type: integer
          format: int64
        blobIds:
          type: array
          items:
            type: string
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ListDaemonVersionsResponse:
      type: object
      properties:
        versions:
          type: array
          items:
            $ref: "#/components/schemas/DaemonVersionSummary"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-platform-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
info:
  title: qits-artifacts API
  version: 2026.922.220434